Task 11 - Refactoring: Changing a Method's Signature.
Changed the signature of a method to Cell.java into a Boolean return
type with parameter named msg of type string.  This created errors to
the code.  In order to fix the errors return types needed to be added to
several areas.  After adding the return type I ran the JUnit test and it
returned no errors.